lightrag-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
lightrag-hku==1.3.8
google-genai==1.14.0
networkx==3.4.2
numpy==1.26.4
